requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
Django==3.2.12
django-admin-interface==0.19.0
django-colorfield==0.6.3
django-filter==21.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-rest-knox==4.2.0
django-rest-passwordreset==1.2.1
djangorestframework==3.12.4
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.1.1
packaging==21.3
Pillow==9.0.1
psycopg2==2.9.3
pycparser==2.21
pyparsing==3.0.7
pytz==2022.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
